MakeCatalog uses geometric when flux weighted not available
Until now (since the re-write, not in version 0.2) when the object fully
lies over a blank region, or is below the threshold, its `--x' and `--y' or
`--ra' and `--dec' would be NaN and its area would be zero. Indeed, that is
the definition of a flux weighted center. But these coordinates are used
generically most of the time and forcing the user to call two sets of
position columns to make a generic catalog is not really convenient.

So now, if the flux weighted center of the object is not available. All
flux-weighted position columns will use the geometric center calculations
instead. Also, the `--area' column now reports the raw area of the pixels
labeled with this object/clump, it is independent of the pixel values they
lie over. To help distinguish if the flux weighte positions were used or
the geometric positions, MakeCatalog now has a new `--weightarea' column.