Fix memory leak in mkcatalog
Memory allocated by `asprintf' should be freed. Most of the similar
allocations are added to list using `gal_list_str_add' and then memory is
freed when the list is deleted. But `tstr' is a temporary variable and
should be freed manually after use.

Assigning NULL to the variable is not necessary but I did it just in case
the variable will be used later in this function.

This fixes bug #51118.