MakeCatalog first pass coordinate increment in shift loop
The loop that calculates the shift of coordinates for estimating
second-order parameters was suppose to also increment the coordinate by one
because in the FITS standard counting starts from one not zero. This was
correctly done in the second pass but it was forgotten in the first and we
were manually incrementing it every time we needed it!

The problem was found because in flux weighted first order positions, the
increment wasn't in a parenthesis. So the final coordinate found would be
wrong.