MakeCatalog brightness is NaN when no pixels were used
In MakeCatalog, until now, when no image pixels were used in measuring the
brightness, the brightness columns would report 0. But 0 is meaningful (it
might indeed be the sum of the pixels of a profile after sky
subtraction). So to avoid such confusions from now on, when no pixels were
used, the reported value is NaN.

Also, to be more similar to the options, the `*_NUMPOS' and `*_SUMPOS'
constants are now renamed to `*_NUMWHT' and `*_SUMWHT'.