NEWS file and webpage updated for 0.3 release
For the 0.3 release, all the recent changes have been incorporated into the
NEWS file. For the new and changed features of the programs, issues related
to one program now start with that program's name followed by a colon. So
if the users can more easily find changes. All the issues are also sorted
by program name.

In the webpage, a new "Installation" title was added which now gives a fast
description of how to install Gnuastro and its mandatory dependencies.