Spell-check in book
Before the 0.3 release, I had forgot to run a spell-check! Emacs's `ispell'
was ran over the whole book and many typos were corrected. I also included
spell-checking (both in the book and in the announcement) in the list of
things to do before distribution.

Also, the Fits and BuildProgram executable files were not in the list of
files to ignore by Git. So they are added.