MakeProfiles deletes merged image if it already exists
By default, MakeProfiles (like all Gnuastro programs), must delete its
outputs if they already exist. However, MakeProfiles was not doing that for
its merged images. It was doing it for the individual profiles, but not the
merged image. With this commit, the merged image file is also deleted if it
already exists.