Corrected NoiseChisel's NUMLABS in onlydetection mode
The value of the `NUMLABS' keyword in NoiseChisel's detected object image
was based on `p->numobjects'. However, this variable was given a value only
after segmentation. So when NoiseChisel was run in `--onlydetection' mode,
the value of this keyword would be `0'! This is now corrected.

Also, to facilitate access to the library, a `libgnuastro' title was also
added to the Info contents. So by running `info libgnuastro' you will be
directly taken to the `gnuastro library' section of the manual.