Warning (and blank output) for quantile with size of zero
When the size of an input dataset is 0 (can happen in automatic
conditions), the quantile is not defined. With this commit, in such cases,
the `gal_statistics_quantile_index' function will print a warning and
`gal_statistics_quantile' (which will call `gal_statistics_quantile_index'
and thus print its warning) will return a blank value of the given dataset.