Corrections for 32-bit systems and in manual
Fortunately GNU Guix was quick to update its Gnuastro image. There were a
few compiler warnings in its log file that are corrected with this
commit. Unfortunately, it still fails on two tests:
`cosmiccal/simpletest.sh' and `mkcatalog/simple.sh'! I have sent a mail to
the Guix developers to see how we can access the `tests/test-suite.log'
file and find the cause of the problem.

In the process, some parts of the manual were also edited to correct for
typos and hard to read parts.