BZERO and BSCALE used to determine type of input
According to the FITS standard, for images, BITPIX doesn't have any code
for `int8_t', `uint16_t', `uint32_t' and `uint64_t' types. So the FITS
standard allows for the `BZERO' and `BSCALE' keywords to be used to create
these types and CFITSIO does the proper conversion. But until now, we
hadn't taken these two keywords into account.

So for example, when `BITPIX=16' (FITS code for `int16'), `BZERO=32768' and
`BSCALE=1', the dataset is actually `uint16'. However, until this commit a
signed 16-bit image would be created to store the values. Hence, if the
image contained large values, we would get a CFITSIO numerical overflow
error (CFITSIO error code 412).

With this commit, the function that reads the image information (including
its numerical data type: `gal_fits_img_info') now also looks for `BZERO'
and `BSCALE' and if they are given and correspond to the standard type
conversions described above, the type of the image is corrected
appropriately. Since it involved reading keywords, to speed things up, it
also reads the name and units if given and so it takes two extra arguments.

In the process of playing with the `BZERO' and `BSCALE', Gnuastro can now
also write `uint64_t' datasets in FITS.

This issue was reported by David Valls-Gaboud.

Improvements:

  - Programs can write unsigned 64-bit integer FITS files.

Changes:

  - `gal_fits_img_info' now also returns the name and units of the dataset
    (if they aren't NULL). So it takes two extra arguments.