Fixed 32-bit check errors and freeing empty WCS
After Gnuastro 0.3.7 was uploaded to the Debian server and tested on many
systems, it was found that two tests fail (one in CosmicCalculator) and
another in MakeCatalog. So after installing an i386 Debian as a virtual
machine and testing it, the causes of the problem were found and corrected:

 - The CosmicCalculator test was corrected by defining a `sum' variable to
   be the sum of the different densities and using that for a check. I
   guess this is a compiler issue. But using that variable, also helped in
   several places where this sum was necessary in the function. Also, since
   users might use scripts to generate the densities, we now allow for
   small floating point errors also.

 - The problem in MakeCatalog was due to not being able to read the
   `NUMLABS' keyword in the clumps image. The problem was that on 32-bit
   systems, `long' and `int' have the same length of 32-bits. But the
   CFITSIO manual is silent on the length of its `TINT' (or `TUINT')
   lengths. So on a 32-bit system, when trying to read a number into
   `TUINT' it would give an overflow error. As a result, when converting
   Gnuastro's types to CFITSIO's types, for 32-bit integers, CFITSIO's LONG
   types are preferred.

Also with this commit, the work done in `gal_wcs_read_fitsptr' of the last
commit was cleaned and some bugs in it were corrected, for example the
`altlin' variable is now only checked when a WCS structure was actually
read.