Type scaling/correction only done when necessary
In commit ea821eb (`BZERO and BSCALE used to determine type of input'), we
applied a correction to the expected type of the dataset when the `BZERO'
and `BSCALE' were present. But they might also be set (for example to 0 and
1 respectively). In such cases, no correction must be done.

From this commit onward, when the scale is `1' and none of the standard
`BZERO' values are given for the specific type, this function won't
actually do anything. Since that commit, it would abort with an error of
unrecognized data type for example when the image was a `float32' type with
`BZERO=0' and `BSCALE=1'.