Freeing of binary arithmetic operators corrected
When the processed output type is different from the type that must be
returned, the binary operators (that need compilation), would change the
type (and thus pointers) of the output dataset before cleaning the
(possibly) temporary left and right operands. This would cause problems
when the output array is changed because when the operation is done in
place, the output dataset is used for the left and right operands and we
would thus try to free the datasets two times!

To solve the problem, changing the final type of the output is done after
(possibly) cleaning the left and right operands.