New --globalhdu (-g) option for Arithmetic
When many files are input into Arithmetic, it may be the case that same HDU
string should be used for all of them. It is thus very inconvenient to call
something like `--hdu=1' 100 times (for 100 input files) for example, even
in a configuration file.

With this commit, Arithmetic now has a new `--globalhdu' (`-g')
option. When this option is called, the `--hdu' option is ignored and for
any input image, Arithmetic will use the string/number given to this
option.

This was suggested by David Valls-Gabaud.