Arrays properly used in Arithmetic's multi-operand operators
The multi-operand operators of `lib/arithmetic' and thus the Arithmetic
program were not manipulating the array elements properly! Instead of
incrementing the pixel indexs, they were incrementing the array
pointers. This would cause a segmentation fault in all these operators.

With this commit, a new `j' counter has been introduced for the pixel
positions and instead of using pointer arithmetic, we just use it to get to
the desired pixel to avoid such confusions.

On a side-note, I noticed that in the `tests/mkcatalog/aperturephot.sh'
script, we were asking for `--ids'. Since its input image is just an
objects image and no clumps are associated with it, this would cause
warnings which can make the output `tests/test-suite.log' file hard to
read. So that column request was removed and this script now just asks for
`--objid'.

The Arithmetic bug was reported by David Valls-Gabaud.