MakeProfiles builds kernel without a catalog with --kernel
Until now, MakeProfiles needed a catalog as input for profile parameters,
the reason is that multiple parameters are needed to define a generic
profile. But a PSF, or convolution kernel, isn't a generic profile and is
highly constrained: for example its center is fixed (the center of the
central pixel), the image size is also dependent on the truncation radius,
and its total sum of values has to be one. So, the only parameters
necessary are the radial function, its FWHM, and its truncation radius.

Making a kernel (for convolution) is necessary for several Gnuastro
programs and it is inconvenient to make a catalog every time you need a
single kernel. So with this commit a `--kernel' option is added to
MakeProfiles. With this option, the kernel parameters can be defined on the
command-line, so there is no need to define a catalog. This greatly
simplifies the creation of kernels/PSFs and testing their effect on the
analysis steps that need it.

This finishes task #14421.