Corrections for 32-bit big endian systems
As part of the Debian checks for Gnuastro version 0.3.13 it was found that
`make check' fails on 32-bit big-endian systems (`mips', `hppa',
`powerpc'). This was reported in Debian bug #865973 and Adrian Bunk was
able to find the cause of the bug (many thanks to Adrian).

To solve the problem, a strategy similar to `GAL_TYPE_SIZE_T' was taken, at
configure time, the width of `long' on the system is found and in
`gnuastro/types.h', we use that to determine which fixed-width type `long'
corresponds to and put that in a `GAL_TYPE_LONG'. This type was then used
for MakeProfile's `--naxis' and `--shift' options.

In a small and separate issue, I recognized that we are not initializing
`status' in Crop's `onecrop_make_array'. With this commit, it is
initialized with a value of zero.

This fixes bug #51341.