Minor cleanups in top webpage and tests/Makefile.am
In the list of dependencies in the top Gnuastro webpage, first we had
CFITSIO, then WCSLIB and then GSL. This was not the same order as the
manual (first GSL, then CFITSIO, then WCSLIB). This caused some confusion
in early users, since the webpage also have a next and previous. So to keep
things consistent, the webpage now has the same order as the manual.

I also noticed that make clean is not removing the `simpleio' file that is
created during `make check'. This file is no included in the `CLEANFILES'
target of `tests/Makefile.am'.