Pure rotation in Warp around pixel coordinate (0,0)
When doing rotation about pixel coordinate (0,0) by an angle theta, until
now Warp would use the following wrong equation (which is correct for
rotation of the coordinates, not rotation of the point):

  cos(theta)  sin(theta)
 -sin(theta)  cos(theta)

When rotating the point (not the coordinates), the negative sign changes on
the `sin' functions must change. However, this was not noticed! The reason
was that in the standard aligned images, the first axis (RA) of the WCS
points to the left, not the right. So the rotation would actually occur not
around image coordinate (0,0), but image coordinate (X,0), where X is the
length of the image along the first axis. This lead us to incorrectly
associate rotation in the WCS axis directions (only when the RA and image X
axis are in opposite directions).

The problem became clear when trying rotation on an image where the RA and
X axis were in the same direction. With this commit, the rotation matrix
calculation is now corrected so like all warpings, the image coordinates
are used, not WCS. The WCS information is just corrected.

While working on this, a small issue in the `--align' option was noticed:
when the first image axis and RA were in the same direction, `--align'
wouldn't do any change! This has also been corrected with this commit.

Also, as part of the fix to the `--align' option, it can now account for
the possiblity that the pixels are not square (different pixel scales on
different dimensions).

This fixes bug #51353.