Dimension-agnostic option names in MakeProfiles
The options of MakeProfiles were too particular to dimensionality and WCS
nature of a dataset, for example `--naxis1', `--xshift' and
`--racol'. These just made too many options and adding different types of
data would be complicated (for example in a spectrum which is a 2D image,
we have length and wavelength as WCS). So `--ra' and `--dec' were
meaningless, but using them was the only way.

With this commit, all MakeProfiles options that had similar problems have
been changed as discussed below. Also all the main WCS parameters can be
defined with options now.

  The PC matrix, CUNIT and CTYPE world coordinate system keywords of the
  output can be set with `--pc', `--cunit' and `--ctype' options. Also the
  old `--crpixN', `--crvalN' and `--resolution' options were replaced with
  `--crpix', `--crval' and `--cdelt' options which can take multiple
  values.

  The new `--naxis' and `--shift' options can take multiple values for each
  dimension (separated by a comma). This replaces the old `--naxis1',
  `--naxis2' and `--xshift' and `--yshift' options.

  The new `--ccol' option can take the center coordinate columns of the
  catalog (in multiple calls) and the new `--mode' option is used to
  identify what standard to interpret them in (image or WCS). Together,
  these replace the old `--xcol', `--ycol', `--racol' and `--deccol'.