NoiseChisel segmentation deals with no usable pixels
It may happen that the number of usable pixels to define clumps over
un-detected pixels in some tiles are zero. However, the sky clump
identification and S/N calculations weren't ready for that. In such a
cases, at the start of segmentation, NoiseChisel would crash with a
"segmentation fault (core dumped)".

The particular case that this happend was when a large area of the
surroundings of the image were blank. So many tiles were blank. However,
the particular image and tile sizes used resulted in some tiles only having
one pixel row of overlap with the non-blank image. Since all pixels on the
edge of a tile are not considered in finding clumps over the sky region,
the result was that there were actually no usable pixels for sky
segmentation within the tile. Such situations won't cause a segmentation
fault any more.

This bug was reported by Raúl Infante Sainz and Nacho Trujillo.

In parallel some corrections to the book were also made regarding the
recent updates to MakeProfiles of the last commit.

This fixes bug #51372.