gal_wcs_pixel_scale returns correct order
The `gal_wcs_pixel_scale' function uses Singular value decomposition (SVD)
to find the pixel scale from the full WCS matrix independent of its
rotation. However, GSL's SVD implementation orders the singluar values
based on their values. It does not preserve the original ordering of the
input matrix.

Until now, `gal_wcs_pixel_scale' would just return the output of GSL's SVD
function. So it was correct when the pixel scales were identical in all
dimensions or when the scales decrease for higher dimensions.

With this commit, the V matrix output of GSL's SVD function is used to find
the correct order and permute the pixel scales correspondingly.

This fixes bug #51385.