Unified numeric type for MakeProfiles function codes
Gnuastro 0.3.23 fails on big-endian systems in the Debian builds (you can
click on the red "Build attempted" signs to see the log of the whole
build. This is the error:

bin/mkprof/.libs/astmkprof: ../tests/mkprof/mkprofcat2.txt: table row 1,
the function code is 0. It should be >0 and <7. Please run again with
`--help' and check the acceptable codes.

The reason of the problem was that we were treating the function codes as a
8-bit integers in some places, and 32-bit integers in others. Since there
is only 6 profiles currently (profile codes between 1 and 7), only the
smallest byte was actually being filled.

On little-endian systems (most commonly used systems including mine), the
smallest byte is placed first, so reading a 32-bit integer as an 8-bit
integer worked successfully and there was no problem. But on big-endian
systems, only the last byte is filled and so the first byte was being read
as 0.

To fix the issue, all the cases we deal with function codes have been
unified to `uint8_t' with this commit (allowing a maximum of 255 function
types).

This fixes bug #51467.