Correct type when storing value of mode in MakeProfiles
the `mode' element of the `mkprofparams' (defined in MakeProfile's
`main.h') was defined as `uint8_t', but in `ui_parse_coordinate_mode', it
was being read as an `int'. So it would crash on many systems (older
compilers).

This was found thanks to crash reports of Debian for Gnuastro 0.3.31.

This fixes bug #51476.