MakeProfiles checks for reasonable values in columns
Until now, there were no checks to see if the column values are reasonable
or not (for example radius is not negative). Tests have been added for this
sanity check on the radius, axis ratio and truncation columns with this
commit.

Also, internally the position columns are defined before the other columns
to be more clear.