Arithmetic works with only one operand
Until now, when Arithmetic was run with only one image and no operators it
would crash. This was because of the behavior we have defined for
efficiency: an image is only actually read into memory when it is ready to
be operated on by an operator, not when it is confronted. In this way, when
processing a large number of files, at any moment only the ones that are
actually being operated on are in memory.

Cases like this (when there is only one file operand and no operator) can
happen in scripts. So when there was no operator, the contents of the image
were not read into memory and we got a segmentation fault. To fix the
problem a check was added after all the input tokens (operators and
operands) were read and applied, so if the dataset is NULL, it reads it
into memory.

This fixes bug #51559.