Correct reading of BZERO for unsigned 64-bit integers
According to FITS standard (v3), "The [BZERO] value field shall contain a
floating-point number ...". Hence, in Gnuastro, we ask CFITSIO to read it
as double precision floating point.

But a double precision floating point can only safely convert an integer to
a float and back to an integer upto 15 significant decimal digits. Hence,
when using BZERO to read the array as a differently signed integer of the
same width, we will get accurate results for signed 8-bit, unsigned 16-bit
and unsigned 32-bit integer datasets.

However, the integer `9223372036854775808' that must be used for BZERO to
read the dataset as an unsigned 64-bit integer has 19 significant decimal
digits. So it (or integers close to it) cannot be accurately resolved in
`double' type (to compare with the standard value).

After consulting with William Pence (author of CFITSIO), he made a great
suggestion, to read the BZERO value as a string and in the case of 64-bit
integers, use the string for comparison, not the floating point. With this
commit, this is implemented.

This fixes bug #51555.