Corrections in tessellation and convolution libraries
The spatial convolution library now uses the higher-level
`GAL_TILE_PO_OISET' macro (the same as `GAL_TILE_PARSE_OPERATE', but with
specific types to be faster) instead of using the low-level tile parsing
features. This makes it much more cleaner to read by eye.

Also, a bug was found in the coordinate incrementation of
`gal_tile_block_increment': we would intentionally set the coord[1] to zero
after the second dimension had been fully parsed. This was only correct
when the coordinate started from zero. It would give a wrong result in
other cases.