New --mcolisbrightness option for MakeProfiles
It might happen that the user knows the total brightness (sum of pixel
values) of the profiles desired, but doesn't want to go through the trouble
of converting it to a magnitude (which also requires a zeropoint). With
this commit, the new `--mcolisbrightness' option will tell MakeProfiles to
interpret the value in `--mcol' as total brightness, not magnitude.