Final noise sigma as input in MakeNoise
Until now, it was necessary to calculate the background magnitude for the
noise-level you wanted. This is indeed the most accurate way to estimate
the noise (since noise depends on the background and the pixel
values). However, when users are just interested in the low surface
brightness regions and (where the noise levels don't change much), it is
much more convenient to just input the noise sigma instead of going through
all the trouble of converting it to a background magnitude with a certain
zeropoint.

Also with this commit, the old `--stdadd' (`-s') option is now called
`--instrumental' (`-i') to be more clear.