Removed bootstrap files from distribution tarball
Bootstrapping is necessary for operations on the version controlled files
and in order to make the tarball. Once the tarball of a specific version of
Gnuastro is released, it is not good to bootstrap it again (for example
download updated Gnulib source files and etc). Changes in the bootstrapping
tools may cause changes in the behavior of the produced Gnuastro programs.

So the `bootstrap', `bootstrap.conf' and `bootstrapped/m4/gnulib-cache.m4'
have been removed from the list of files that are to be distributed in
future tarballs.