Convolve name and dimensionality checks
The Convolve program wasn't checking the sanity of the input file names and
the dimensionality of the input. These checks are now implemented in
`bin/convolve/ui.c'.

Also to help in reading, the spaces in `gal_fits_name_is_fits' were
adjusted and `gal_fits_suffix_is_fits' was made more simple/efficient by
avoiding simultaneous checks with and without a dot.