Corrections in NoiseChisel check extension names
NoiseChisel's check image extension names have a `-' between words when a
multi-word name is necessary. However, `DETECTION_FINAL' used an `_'. This
has been corrected. Two other extension names were changed.
`INPUT-CONVOLVED' was changed to `CONVOLVED' because the `INPUT-' part was
irrelevant. `TRUE-PSEUDO-DETECTIONS' was changed to `TRUE-PSEUDOS' to be
shorter (easier to type) and fit better with the other occurrences of
`PSEUDOS' in the extension names of the `--checkdetection' output
(`PSEUDOS-FOR-SN' which occurs two times).

Also, the problem that the previous commit solves (reading CRLF line
terminated ASCII text files) was raised by Marjan Akbari. But I forgot to
add her name to the `THANKS' file, it is now done with this commit.