Pushing tags in Debian release notes
In the Debian release checklist, we weren't pushing the created tags onto
the repo (informed by Ole Streicher). So a `--tags' was added to the final
`git push' command. Also, `origin' was removed, because it is irrelevant.

Also, the NEWS file was updated to say that all programs can now benefit
from the reading of CRLF line terminator ASCII text files. From the
previous version of the NEWS file, normal user might think that it is only
possible in the libraries.