Suggestion on inspecting NoiseChisel's output
In the "NoiseChisel output" section of the book, we had discussed the
different extensions. But I had forgot to refer the readers to the appendix
which describes how best to view them. So a short paragraph was added.

Also, the node name of the "Vewing multiextension FITS images" appendix
section was set to be the same as the section name.