Write access only when necessary in Fits program
Until now, when the Fits program was asked to copy an extension to another
or cut it, it would open the FITS file in `READWRITE' mode. This wouldn't
work when the input was a `fits.gz' file. So now, only when the user asks
to cut an extension, does `fits_hdu_copy' ask for write access. When the
user just wants to copy a HDU to another file, it will ask for `READONLY'
permissions.