Crop includes BUNIT in output files if present
Until now, the Crop program didn't care about the units of the data and
just copied the dataset and its WCS information. But when `BUNIT' is
present in the input's keywords, it is something like WCS and will probably
matter for the user. Also, Crop doesn't change anything about the dataset,
so the units remain unchanged. Therefore with this commit, when Crop is
creating the output crop, it also checks for the `BUNIT' keyword and if
present, it will write it in the output cropped image also.