FITS keyword titles aligned
The "Versions and date" title of the FITS keywords was not aligned with the
WCS title or the Crop titles. To fix this in any place that a keyword title
is necessary, a macro was defined in `lib/gnuastro/fits.h' which is now
used everywhere.

Also, there was one extra blank line before the "Crop information" title in
Crop's outputs which is now corrected.