Only program names for FITS image keyword titles
Until now, the title of the set of FITS image keywords releated to a
special program was set with the `PROGRAM_STRING' macro. This macro has the
format of "ProgramName (GNU Astronomy Utilities) X.Y.ZZZ". However, in FITS
images, we have a special `GNUASTRO' keyword that has the version number as
value. So the extra information was redundant/repeated.

With this commit, all the cases where an image is written to a FITS file,
the string is changed to `PROGRAM_NAME' which is just the name of the
program.