Elaborated error messages for incorrect HDU
When the input FITS file only has one extension (and the default value to
`--hdu' is 1 to read the second extension) Gnuastro aborts with a
program. But the default CFITSIO error can confuse the (especially early)
users and isn't clear enough. So a much more elaborate error message is now
printed in such cases. It explains the problem thoroughly (in a footnote)
and also suggests the fix to the problem (when the input HDU was 1).

Also, a more clear error message was prepared when the given HDU generally
doesn't exist in the input FITS file. It now suggests using Gnuastro's Fits
program to get a list of the HDUs.

This issue was raised by Fernando Buitrago.