Checks on the output's directory
Until now, if the input to the old `gal_checkset_check_remove_file' didn't
exist (the file, not necessarily the directory) no error was
reported. However, when the user doen't have write permissions in the
directory or it doesn't exist, later errors (usually CFITSIO) may not be
clear enough. So if the file doesn't exist, this file now checks its
directory also and aborts if it is not suitable. Also, the name of the
function was changed to `gal_checkset_writable_remove'.

This process actually started by noticing that in the Fits program, we were
mistakenly feeding this function with the input filename, not the output
filename! So this tiny bug is also corrected with this commit.