More clear "Installation directory" section in book
The Installation directory section of the book was edited to be more
clear. In particular, the `echo " ... " > .bashrc' commands were removed
and only the line that must be put in the startup file is now shown. Also,
to be easy, a summary box/cartouche was added that has all the necessary
lines for the user to simply copy and paste.