Correct index array for clumps when checking S/N table
In `clumps_true_find_sn_thresh', we have to use `GAL_TYPE_INT32' directly
instead of relying on `clprm.snind->type'. This is because of a recent
correction in `clumps_make_sn_table', where we stop continuing the function
when `cltprm->numinitclumps==0' (so `clprm.snind->type' was not initialized
and remained as zero which is an un-recognized type code).