Dilation of true detections replaced with growth
Until now, dilation was used to grow the true detections (because we
initially carved layers of pixels off). But dilation is a blind process, it
doesn't follow the signal, it is just based on the shape of the initial
detected region. As a result to go down to low surface brightness, the
objects would get boxy (if 8-connectivity was used) or diamond shaped (if
4-connectivity was used).

With this commit, a process like the growth of clumps (to define objects)
is used. It is indeed slower than dilation, but it is able to identify the
shape of very low surface brightness signal accurately. See the manual for
a more detailed discussion.

In the libraries, it is now possible to define the connectivity of the
holes to fill in the `gal_binary_fill_holes' function.