NoiseChisel timings corrected
In some places during detected, NoiseChisel would not calculate the current
time, so when reporting the time, the time appeared cumulative. This is now
corrected and the time steps for each step are correctly reported. Also in
the end of the detection process, it says how long the whole detection
process took.