Removed extra freeing when changing values in ConvertType
To change values, ConvertType uses the arithmetic library. The flags that
are passed to the library include `GAL_ARITHMETIC_FREE'. So any temporary
arrays are freed by the library. However, we were mistakenly freeing one
temporary array (`cond') separately afterwards. This resulted in a
segmentation fault. So the freeing of `cond' is now removed.

This bug was found thanks to Takashi Ichikawa.

This fixes bug #52010.