Integer blank values in arithmetic and the FITS BLANK keyword
This commit solves some issues related to integer blank values in the
Arithmetic library and the BLANK keyword that must be written to the FITS
headers. They were both found while trying to do a simple arithmetic
operation on an unsigned 16-bit image and noticing that blank values are
not being considered properly in both stages.

For reading and writing a FITS image, we now have the
`gal_fits_key_img_blank' function which will allocate and fill the memory
for Gnuastro's blank values of any type as required by the FITS standard
(because the BLANK value must be the value before BZERO and BSCALE are
applied).

In the process, I also noticed the integer limits reported in the manual
for 8-bit and 16-bit signed integers needed corrections which are now
applied.

Also, in preparation for version 0.5, the library version has been
incremented to 3.0.0.

This fixes bug #52014.