Wide kernel for NoiseChisel's qthresh estimation
NoiseChisel uses the difference between the mode and median to identify if
a tile should be used for estimating the quantile thresholds. Until now,
NoiseChisel would convolve an image once and estimate the proper tiles for
quantile estimations on the convolved image and use the same convolved
image for quantile estimation. While a larger kernel does increase the
skewness (and thus difference between the mode and median), it disfigures
the shapes of the objects.

With this commit, there is a new `--widekernel' (and a corresponding
`--wkhdu' option to specify its HDU) option. When its given, the image will
be convolved with both the sharp and wide kernels. The mode and median are
calculated on the dataset that is convolved with the wider kernel, then the
quantiles are estimated on the image convolved with the sharper kernel.

Library functions that are changed with this commit:

 - `gal_binary_fill_holes' now accepts a `maxsize' argument to only fill
   holes less than a given size.

 - `gal_tile_block_write_const_value' and `gal_tile_full_values_write' now
   accept a new `withblank' option to set all pixels that are blank in the
   tile's block to be blank in the check image.