NoiseChisel issues with memory mapping fixed
When running NoiseChisel with `--minmapsize=0' (to not use the RAM at all
for arrays), a segmentation fault occured. It was because of changing the
array pointer in one place without correctly changing the memory mapped
filename. This issue is now corrected with this commit.

ALso, I noticed that in `gal_data_free_contents', we weren't setting block
to NULL and `mmapname' to NULL (after freeing it).

Finally, an easier way to ask `./tmpfs-config-make' to compile in debugging
mode was defined. Until now, it was necessary to set/remove a comment close
to the end of the line. But now, through a variable (`debug_noshared') at
the start of the file, this behavior can easily be set.